fix: age groups types

+++ etl/load/models/age_groups.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String
 
 from etl.load.db import Base
 
@@ -8,5 +8,5 @@ class AgeGroup(Base):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String(255), nullable=False)
-    range_start = Column(Float, nullable=True)
-    range_end = Column(Float, nullable=True)
+    range_start = Column(Integer, nullable=True)
+    range_end = Column(Integer, nullable=True)
